requirements.txt
Flask==2.0.2
gunicorn==20.1.0
google-cloud-datastore==2.18.0
